requirements.txt
opencv-python
cmake
pyyaml
streamlit==1.22.0
dlib==19.22.1
numpy
pydub

#Fork of face_recognition using prebuild dlib library instead
face_recognition @ git+https://github.com/thetoby9944/face_recognition



